api/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
numpy==1.26.1
pandas==2.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
